useSocket in shared folder
